Add trace step forward button
Cursor agent prompt:

In the main UI, I want to add a "Forward" button, so that when a user is exploring a trace, and then click "Back", they can also click "Forward" in order to move back forward through the trace they just stepped back in. This should be doable by maintaining a stack of the last states that were stepped back through.