Generalize interpreter handling of bounded quantifier expressions and \in, \notin expressions
This also allows the disablement of two key syntactic rewrites that are no longer required for correctness of evaluation,
including the rewriting of bounded quantifiers and of \in/ or \notin expressions. Now these should all just be handled natively
during evaluation. Eventual goal here is to move away from any syntactic rewriting at all, so that source mapping for eval errors
is greatly simplified.